refactor: :recycle: Better modules separation
