+++ boosty_downloader/src/yaml_configuration/sample_config.py
